+++ src/commonsos/service/command/CreateCommunityCommand.java
@@ -1,5 +1,6 @@
 package commonsos.service.command;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Getter;
@@ -13,7 +14,7 @@ public class CreateCommunityCommand {
   private String communityName;
   private String tokenName;
   private String tokenSymbol;
-  private Double transactionFee;
+  private BigDecimal transactionFee;
   private String description;
   private List<Long> adminIdList;
   private UploadPhotoCommand uploadPhotoCommand;
