+++ src/commonsos/service/command/CreateRedistributionCommand.java
@@ -1,5 +1,7 @@
 package commonsos.service.command;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
@@ -9,5 +11,5 @@ public class CreateRedistributionCommand {
   private Long communityId;
   private boolean isAll;
   private Long userId;
-  private Double redistributionRate;
+  private BigDecimal redistributionRate;
 }
